Fixed coding error. strlen() returns size_t causing a size_t sized value pushed onto stack instead of long. On x64 builds, size_t != long. Caught via compiler warning in x64 build.

+++ util/dlb_main.c
@@ -506,7 +506,7 @@ long slen, dir_size, flen;
             (long) DLB_VERS,   /* version of dlb file */
             (long) nfiles + 1, /* # of entries (includes directory) */
                                /* string length + room for nulls */
-            (long) slen + strlen(DLB_DIRECTORY) + nfiles + 1,
+            (long) slen + (long) strlen(DLB_DIRECTORY) + nfiles + 1,
             (long) dir_size,         /* start of first file */
             (long) flen + dir_size); /* total file size */
     Write(out, buf, strlen(buf));
